+++ changelog.md
@@ -4,6 +4,7 @@
 - Support for Bot API 3.6.
 - Support for multipart attachments.
 - Add methods `Message::getCaptionHtml()` and `Message::hasHtmlCaption()`.
+- Add method `TelegramResponseException::retryAfter()`.
 
 # Version 0.5.0